+++ src/systems/singularity/cinttamobi/util/VEMTrabalhador.java
@@ -12,14 +12,14 @@ public class VEMTrabalhador extends VEM {
     }
 
     @Override
-    public void credit(int value) {
+    public void credit(double value) {
         if (value <= 0)
             throw new RuntimeException("Valor inválido");
         this.balance += value;
     }
 
     @Override
-    public void debit(int value) {
+    public void debit(double value) {
         if (value <= 0)
             throw new RuntimeException("Valor inválido");
         else if (value > this.getBalance())
